# docs/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Evidata.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:65382/"
    }
  ],
  "paths": {
...
    "/api/gaps": {
      "get": {
        "tags": [
          "Gaps"
        ],
        "parameters": [
          {
            "name": "severity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComplianceGapDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComplianceGapDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComplianceGapDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Gaps"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGapRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGapRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceGapDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceGapDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceGapDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "UserProfile"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "UserProfile"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/users/link": {
      "post": {
        "tags": [
          "UserProfile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkExternalIdentityCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkExternalIdentityCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LinkExternalIdentityCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/mcp/query": {
      "post": {
        "tags": [
          "Mcp"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpQueryApiRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/McpQueryApiRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/McpQueryApiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/McpQueryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpQueryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpQueryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mcp/interactions": {
      "get": {
        "tags": [
          "Mcp"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "riskLevel",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/McpInteractionHistory"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpInteractionHistory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpInteractionHistory"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mcp/interactions/{interactionId}/feedback": {
      "post": {
        "tags": [
          "Mcp"
        ],
        "parameters": [
          {
            "name": "interactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/McpFeedbackRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/McpFeedbackRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/McpFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/mcp/interactions/{interactionId}/verify-citations": {
      "post": {
        "tags": [
          "Mcp"
        ],
        "parameters": [
          {
            "name": "interactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CitationVerificationReport"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationVerificationReport"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationVerificationReport"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/roles/users/{userId}": {
      "get": {
        "tags": [
          "Roles"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tenantId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/roles/assign": {
      "post": {
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleToUserCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/roles/remove": {
      "delete": {
        "tags": [
          "Roles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoleFromUserCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignRoleToUserCommand": {
        "required": [
          "userId",
          "roleId",
          "tenantId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CitationVerificationReport": {
        "required": [
          "interactionId",
          "results"
        ],
        "type": "object",
        "properties": {
          "interactionId": {
            "type": "string",
            "format": "uuid"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CitationVerificationResult"
            }
          },
          "verifiedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "notFoundCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "unverifiableCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isFullyVerified": {
            "type": "boolean"
          }
        }
      },
      "CitationVerificationResult": {
        "required": [
          "citationId",
          "sourceType",
          "sourceId",
          "status",
          "reason"
        ],
        "type": "object",
        "properties": {
          "citationId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceType": {
            "$ref": "#/components/schemas/McpCitationSourceType"
          },
          "sourceId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/CitationVerificationStatus"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CitationVerificationStatus": {
        "type": "integer"
      },
      "ComplianceGapDto": {
        "required": [
          "id",
          "tenantId",
          "sourceModule",
          "sourceEntityId",
          "title",
          "description",
          "severity",
          "status",
          "blocksApproval",
          "ownerId",
          "dueAt",
          "createdBy",
          "createdAt",
          "lastModifiedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceModule": {
            "type": "string"
          },
          "sourceEntityId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "blocksApproval": {
            "type": "boolean"
          },
          "ownerId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "dueAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "CreateGapRequest": {
        "required": [
          "sourceModule",
          "sourceEntityId",
          "title",
          "description"
        ],
        "type": "object",
        "properties": {
          "sourceModule": {
            "type": "string"
          },
          "sourceEntityId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "default": "Medium"
          },
          "legalObligationId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "dueAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "LinkExternalIdentityCommand": {
        "required": [
          "externalId",
          "provider",
          "email",
          "displayName",
          "tenantId"
        ],
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "McpCitationSourceType": {
        "type": "integer"
      },
...
      "McpFeedbackRating": {
        "type": "integer"
      },
      "McpFeedbackRequest": {
        "required": [
          "rating",
          "comment"
        ],
        "type": "object",
        "properties": {
          "rating": {
            "$ref": "#/components/schemas/McpFeedbackRating"
          },
          "comment": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "McpInteractionEntry": {
        "required": [
          "interactionId",
          "userId",
          "question",
          "riskLevel",
          "status",
          "usedTenantContext",
          "requiresHumanReview",
          "citationCount",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "interactionId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "question": {
            "type": "string"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/McpRiskLevel"
          },
          "status": {
            "$ref": "#/components/schemas/McpInteractionStatus"
          },
          "usedTenantContext": {
            "type": "boolean"
          },
          "requiresHumanReview": {
            "type": "boolean"
          },
          "citationCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "McpInteractionHistory": {
        "required": [
          "items",
          "totalCount",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/McpInteractionEntry"
            }
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
      "McpInteractionStatus": {
        "type": "integer"
      },
      "McpQueryApiRequest": {
        "required": [
          "question"
        ],
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          }
        }
      },
      "McpQueryResponse": {
        "required": [
          "interactionId",
          "answer",
          "riskLevel",
          "requiresHumanReview",
          "usedTenantContext",
          "ratReferences"
        ],
        "type": "object",
        "properties": {
          "interactionId": {
            "type": "string",
            "format": "uuid"
          },
          "answer": {
            "type": "string"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/McpRiskLevel"
          },
          "requiresHumanReview": {
            "type": "boolean"
          },
          "usedTenantContext": {
            "type": "boolean"
          },
          "ratReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatContextReference"
            }
          },
          "abstentionReason": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "McpRiskLevel": {
        "type": "integer"
      },
...
      "RatContextReference": {
        "required": [
          "activityId",
          "name",
          "department"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "department": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RemoveRoleFromUserCommand": {
        "required": [
          "userId",
          "roleId",
          "tenantId"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "roleId": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}